[Calcul Entity] Création des calculs pour les additions et le QCM
